+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Added `slider_get_set` function to update slider value. Useful for converting to/from integer values.
 * Added utility trait `Mutable<T>` to abstract over `T`, `Mut<T>` types.
 * Added `interaction_enabled` function that does the opposite of `interaction_disabled` (Manages `InteractionDisabled` component)
+* Added fallback locations for anchored elements when outside view.
+* Added `on_hash_change_insert`, `on_hash_change_typ_insert` functions to insert bundle into entity when provided value changes for given key.
 
 # 0.3.1